requirements.txt
# --------- pytorch --------- #
# torch>=1.10.0
# torchvision>=0.11.0
pytorch-lightning>=1.8.1
torchmetrics>=0.10.0

# --------- hydra --------- #
hydra-core>=1.2.0
hydra-colorlog>=1.2.0
hydra-optuna-sweeper>=1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
albumentations
pathlib
matplotlib
opencv-contrib-python
imageio
nibabel
pydicom
scipy
scikit-image
pandas
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
